updates to fix alias bugs
