+++ .vscode/launch.json
@@ -35,7 +35,7 @@
         "--p",
         "sp",
         "--s",
-        "batch",
+        "alias",
         "--site",
         "https://318studios.sharepoint.com/sites/dev/6b6baf3e-e344-41e3-aa42-c47d4ad1be24"
       ]
